move function to common module
